+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2023-12-14
+
+### Added
+
+- AnalyticsManager module
+- PrivacyManager module
+
+## Changed
+- Bump Android Gimbal SDK version 4.9.0 -> 4.9.2
+- Bump iOS Sample App minimum iOS version 12.4 -> 13.0
+- Bump React Native version 0.70.6 -> 0.72.0
+- Sample App metro config and dependencies to accommodate React Native 0.72
+
 ## [0.1.2] - 2022-01-26
 
 ### Fixed
@@ -13,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.1] - 2022-01-23
 
-This is purely an administrative release -- no code changes
+- This is purely an administrative release -- no code changes
 
 ## [0.1.0] - 2022-01-23